requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2

# 환경 변수 관리
python-dotenv==1.0.0

# API 서버
fastapi==0.104.1
uvicorn==0.24.0

# 스케줄링
schedule==1.2.0

# Supabase 클라이언트
supabase==2.0.0
